chore: 패키지 이름 수정으로 인한 변경 파일 업데이트
resolves: #256
