+++ src/main/java/com/zerobase/foodlier/global/refrigerator/controller/RefrigeratorController.java
@@ -6,7 +6,7 @@ import com.zerobase.foodlier.common.response.ListResponse;
 import com.zerobase.foodlier.common.security.provider.dto.MemberAuthDto;
 import com.zerobase.foodlier.module.notification.domain.type.NotificationType;
 import com.zerobase.foodlier.module.notification.dto.notify.NotifyInfoDto;
-import com.zerobase.foodlier.module.notification.dto.notify.impl.ChefNotify;
+import com.zerobase.foodlier.module.notification.dto.notify.Impl.ChefNotify;
 import com.zerobase.foodlier.global.notification.facade.NotificationFacade;
 import com.zerobase.foodlier.global.refrigerator.facade.RefrigeratorFacade;
 import com.zerobase.foodlier.module.member.chef.dto.AroundChefDto;
@@ -16,7 +16,7 @@ import com.zerobase.foodlier.module.member.chef.type.ChefSearchType;
 import com.zerobase.foodlier.module.member.member.dto.RequestedMemberDto;
 import com.zerobase.foodlier.module.member.member.service.MemberService;
 import com.zerobase.foodlier.module.member.member.type.RequestedOrderingType;
-import com.zerobase.foodlier.module.notification.dto.notify.impl.RequesterNotify;
+import com.zerobase.foodlier.module.notification.dto.notify.Impl.RequesterNotify;
 import com.zerobase.foodlier.module.request.dto.RequestDetailDto;
 import com.zerobase.foodlier.module.request.service.RequestService;
 import com.zerobase.foodlier.module.requestform.dto.RequestFormDetailDto;
