+++ src/main/java/com/zerobase/foodlier/module/notification/dto/notify/Impl/RequesterNotify.java
@@ -1,11 +1,10 @@
-package com.zerobase.foodlier.module.notification.dto.notify.impl;
+package com.zerobase.foodlier.module.notification.dto.notify.Impl;
 
-import com.zerobase.foodlier.module.comment.comment.domain.model.Comment;
-import com.zerobase.foodlier.module.heart.domain.model.Heart;
 import com.zerobase.foodlier.module.member.member.domain.model.Member;
 import com.zerobase.foodlier.module.notification.domain.type.NotificationType;
 import com.zerobase.foodlier.module.notification.dto.notify.Notify;
 import com.zerobase.foodlier.module.notification.dto.notify.NotifyInfoDto;
+import com.zerobase.foodlier.module.request.domain.model.Request;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -13,26 +12,27 @@ import lombok.NoArgsConstructor;
 
 import java.util.List;
 
-import static com.zerobase.foodlier.module.notification.constant.MessageConstant.*;
-import static com.zerobase.foodlier.module.notification.constant.MessageConstant.CLOSE_BRACKET;
+import static com.zerobase.foodlier.module.notification.constant.MessageConstant.DELIMITER;
+import static com.zerobase.foodlier.module.notification.constant.MessageConstant.HONORIFIC_TITLE;
 
 @Getter
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 @Builder
-public class CommentNotify implements Notify {
+public class RequesterNotify implements Notify {
+
     private Member receiver;
     private String performerNickname;
     private Long targetSubjectId;
     private String targetTitle;
     private NotifyInfoDto notifyInfoDto;
 
-    public static CommentNotify from(Comment comment, NotifyInfoDto notifyInfoDto){
-        return CommentNotify.builder()
-                .receiver(comment.getRecipe().getMember())
-                .performerNickname(comment.getMember().getNickname())
-                .targetSubjectId(comment.getRecipe().getId())
-                .targetTitle(comment.getRecipe().getSummary().getTitle())
+    public static RequesterNotify from(Request request, NotifyInfoDto notifyInfoDto){
+        return RequesterNotify.builder()
+                .receiver(request.getMember())
+                .performerNickname(request.getChefMember().getMember().getNickname())
+                .targetSubjectId(request.getId())
+                .targetTitle(request.getTitle())
                 .notifyInfoDto(notifyInfoDto)
                 .build();
     }
@@ -47,7 +47,7 @@ public class CommentNotify implements Notify {
 
     @Override
     public NotificationType getNotificationType() {
-        return this.notifyInfoDto.getNotificationType();
+        return notifyInfoDto.getNotificationType();
     }
 
     @Override
